refactor: Rename `lib/removeDestination.js`
