+++ lib/plugin.js
@@ -12,7 +12,7 @@ const injectLogsIamRole = require('./inject-logs-iam-role');
 const injectOutputOutputs = require('./inject-output-outputs');
 const wrapClean = require('./wrapClean');
 const getAppUids = require('./app-uids');
-const removeDestination = require('./removeDestination');
+const removeDestination = require('./remove-destination');
 const { saveDeployment, createAndSetDeploymentUid } = require('./deployment');
 const variables = require('./variables');
 const { generate } = require('./generate-event');
