+++ lib/plugin.test.js
@@ -46,7 +46,7 @@ describe('plugin', () => {
       './logs-collection': logsCollection,
       './wrap': wrap,
       './wrapClean': wrapClean,
-      './removeDestination': removeDestination,
+      './remove-destination': removeDestination,
       './deployment': { saveDeployment, createAndSetDeploymentUid },
       './generate-event': { eventDict: {}, generate },
       './deploy-profile': { configureDeployProfile },
