+++ lib/remove-destination.js
