+++ lib/remove-destination.test.js
@@ -9,7 +9,7 @@ describe('removeDestination', () => {
   let removeDestination;
   before(() => {
     removeLogDestinationSpy = sinon.spy();
-    removeDestination = proxyquire('./removeDestination', {
+    removeDestination = proxyquire('./remove-destination', {
       './client-utils': {
         getPlatformClientWithAccessKey: () => ({
           logDestinations: {
